add LatencyEvent class
